Feature #9: Use GitHub Actions to deploy on vercel instead of the fork way
# Story Title

[Feature: Use GitHub Actions to deploy on vercel instead of the fork way](https://github.com/kuru-studio/kuru-studio-social-web/issues/9)

## Changes made

- Add release.yml

## How does the solution address the problem

This PR will move the deploying of the repository to Vercel from BosEriko's fork to a GitHub action here inside this repository. Please check [this](https://github.com/thimble-bot/web/blob/master/.github/workflows/release.yml) out and be sure to create a skip-ci function. Check out your [Notion](https://www.notion.so/kuru-studio/Use-GitHub-Actions-to-deploy-on-vercel-instead-of-the-fork-way-a05803aebaba454392c4aaa850a2375d) for information on how to do the skip-ci.

## Linked issues

Resolves #9